moving ReactTab component with its css file into its own folder and changing code so that state controls what class ReactTab is rendered with
