+++ src/containers/ReactTab/ReactTab.css
@@ -1,0 +1,28 @@
+#reactTab {
+  padding: 10px;
+  display: grid;
+  grid-template-rows: 1fr 3fr;
+  grid-template-columns: 3fr 7fr;
+  grid-template-areas:
+    'sp-box tree'
+    'diff tree';
+}
+
+#reactLoader {
+  margin: 30px;
+  text-align: center
+}
+
+/* for different screen sizes */
+@media screen and (max-width: 600px) {
+  #reactTab {
+    display: grid;
+    grid-template-rows: 4fr 3fr 4fr;
+    grid-template-columns: 1fr;
+    grid-template-areas:
+      'diff'
+      'sp-box'
+      'tree';
+    max-height: 1200px;
+  }
+}
